requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain-core==0.1.42
langchain-google-genai==1.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
pillow==10.1.0